requirements.txt
# LLM
anthropic>=0.49.0

# Market data
yfinance>=0.2.54
pandas>=2.2.0
numpy>=1.26.0

# News
newsapi-python>=0.2.7

# HTTP (Alpha Vantage + SEC EDGAR use requests directly)
requests>=2.32.0

# Charts
matplotlib>=3.9.0
seaborn>=0.13.0

# Web UI
streamlit>=1.43.0

# Utilities
python-dotenv>=1.0.0
